students page refresh

+++ src/app/pages/lesson/lesson.component.ts
@@ -1,7 +1,10 @@
-import { MatTableDataSource, MatSort } from '@angular/material';
+import { MatTableDataSource, MatSort, MatDialog } from '@angular/material';
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { Lesson } from '../../models/lesson.model';
 import { LessonService } from '../../services/lesson/lesson.graphql.service';
+import { SubscriptionCleaner } from '../../decorators/SubscriptionCleaner.decorator';
+import { first } from 'rxjs/operators';
+import { DeleteClassDialogComponent } from '../class/dialogs/delete/delete-class.dialog';
 
 @Component({
   selector: 'app-lesson',
@@ -15,16 +18,41 @@ export class LessonComponent implements OnInit {
   @ViewChild(MatSort)
   sort: MatSort;
 
-  constructor(private lessonService: LessonService) {}
+  @SubscriptionCleaner()
+  subCollector;
+  constructor(private lessonService: LessonService, public dialog: MatDialog) {}
 
   async ngOnInit() {
-    const lessons = await this.lessonService.getLessons();
-    this.dataSource.data = [...lessons];
+    try {
+      this.subCollector.add(
+        this.lessonService.getAllLessons().subscribe((lessons) => {
+          this.dataSource.data = [...lessons];
+        }),
+      );
+    } catch (error) {
+      // TODO: implement error handling on UI
+      console.error('Error handling not implemented');
+      throw error;
+    }
+    //const lessons = await this.lessonService.getLessons();
+    //this.dataSource.data = [...lessons];
   }
 
   addNewLesson() {}
 
-  deleteLesson(_id: string) {
-    this.lessonService.delete(_id);
+  public async deleteLesson(_id: string) {
+    const dialogRef = this.dialog.open(DeleteClassDialogComponent, {
+      data: { _id, name, level: null },
+    });
+    this.subCollector.add(
+      dialogRef
+        .afterClosed()
+        .pipe(first())
+        .subscribe(async (result) => {
+          if (result === true) {
+            await this.lessonService.delete(_id);
+          }
+        }),
+    );
   }
 }
